Move LayoutAssets to a new namespace
